restructure include folder and tests
