+++ .gitignore
@@ -1,6 +1,3 @@
 .DS_Store
 mason_packages
-test
-node_modules
-test-Release
-test-Release.dSYM
+build
