+++ include/hello_world/version.hpp
@@ -1,0 +1,16 @@
+#pragma once
+
+/// The major version number
+#define VERSION_MAJOR 0
+
+/// The minor version number
+#define VERSION_MINOR 0
+
+/// The patch number
+#define VERSION_PATCH 1
+
+/// The complete version number
+#define VERSION_CODE (VERSION_MAJOR * 10000 + VERSION_MINOR * 100 + VERSION_PATCH)
+
+/// Version number as string
+#define VERSION_STRING "0.0.1"
